feat: move AppUser model and add UserObserverRegistry to user package
