+++ src/main/kotlin/com/safehill/safehillclient/data/authorization/UserAuthorizationRepository.kt
@@ -9,8 +9,8 @@ import com.safehill.kclient.network.WebSocketApi
 import com.safehill.kclient.tasks.syncing.InteractionSync
 import com.safehill.kclient.util.safeApiCall
 import com.safehill.safehillclient.manager.dependencies.UserObserver
-import com.safehill.safehillclient.model.user.AppUser
-import com.safehill.safehillclient.model.user.toAppUser
+import com.safehill.safehillclient.data.user.model.AppUser
+import com.safehill.safehillclient.data.user.model.toAppUser
 import com.safehill.safehillclient.module.client.UserScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
