+++ src/main/kotlin/com/safehill/safehillclient/data/threads/registry/ThreadStateRegistry.kt
@@ -7,7 +7,7 @@ import com.safehill.kclient.models.users.UserProvider
 import com.safehill.kclient.models.users.getOrNull
 import com.safehill.safehillclient.data.threads.model.MutableThreadState
 import com.safehill.safehillclient.data.threads.model.ThreadState
-import com.safehill.safehillclient.model.user.toAppUser
+import com.safehill.safehillclient.data.user.model.toAppUser
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
